+++ DemoApps/Shared/OpenCL/Helper/UnitTest/source/UnitTest/Test_TightPlanarBitmapR8G8B8.cpp
@@ -171,8 +171,8 @@ TEST(Test_TightPlanarBitmapR8G8B8, ConstructFromReadOnlySpan_Extent)
   constexpr uint32_t ExpectedPlaneStride = Stride / 3;
   constexpr uint32_t ExpectedBytesPerPlane = ExpectedPlaneStride * Extent.Height.Value;
 
-  const std::array<uint8_t, 3 * Extent.Width.Value * Extent.Height.Value> srcContent{0x00, 0x40, 0x80, 0x10, 0x50, 0x90, 0x01, 0x41, 0x81,
-                                                                                     0x11, 0x51, 0x91, 0x02, 0x42, 0x82, 0x12, 0x52, 0x92};
+  const std::array<uint8_t, static_cast<std::size_t>(3) * Extent.Width.Value * Extent.Height.Value> srcContent{
+    0x00, 0x40, 0x80, 0x10, 0x50, 0x90, 0x01, 0x41, 0x81, 0x11, 0x51, 0x91, 0x02, 0x42, 0x82, 0x12, 0x52, 0x92};
 
   const TightPlanarBitmapR8G8B8 result(SpanUtil::AsReadOnlySpan(srcContent), Extent, SrcPixelFormat, Origin);
 
@@ -201,8 +201,8 @@ TEST(Test_TightPlanarBitmapR8G8B8, ConstructFromRawBitmap)
   constexpr uint32_t ExpectedPlaneStride = Stride / 3;
   constexpr uint32_t ExpectedBytesPerPlane = ExpectedPlaneStride * Extent.Height.Value;
 
-  const std::array<uint8_t, 3 * Extent.Width.Value * Extent.Height.Value> srcContent{0x00, 0x40, 0x80, 0x10, 0x50, 0x90, 0x01, 0x41, 0x81,
-                                                                                     0x11, 0x51, 0x91, 0x02, 0x42, 0x82, 0x12, 0x52, 0x92};
+  const std::array<uint8_t, static_cast<std::size_t>(3) * Extent.Width.Value * Extent.Height.Value> srcContent{
+    0x00, 0x40, 0x80, 0x10, 0x50, 0x90, 0x01, 0x41, 0x81, 0x11, 0x51, 0x91, 0x02, 0x42, 0x82, 0x12, 0x52, 0x92};
 
   const ReadOnlyRawBitmap srcRawBitmap = ReadOnlyRawBitmap::Create(SpanUtil::AsReadOnlySpan(srcContent), Extent, SrcPixelFormat, Origin);
   const TightPlanarBitmapR8G8B8 result(srcRawBitmap);
@@ -232,8 +232,8 @@ TEST(Test_TightPlanarBitmapR8G8B8, ConstructFromRawBitmapEx)
   constexpr uint32_t ExpectedPlaneStride = Stride / 3;
   constexpr uint32_t ExpectedBytesPerPlane = ExpectedPlaneStride * Extent.Height.Value;
 
-  std::array<uint8_t, 3 * Extent.Width.Value * Extent.Height.Value> srcContent{0x00, 0x40, 0x80, 0x10, 0x50, 0x90, 0x01, 0x41, 0x81,
-                                                                               0x11, 0x51, 0x91, 0x02, 0x42, 0x82, 0x12, 0x52, 0x92};
+  std::array<uint8_t, static_cast<std::size_t>(3) * Extent.Width.Value * Extent.Height.Value> srcContent{
+    0x00, 0x40, 0x80, 0x10, 0x50, 0x90, 0x01, 0x41, 0x81, 0x11, 0x51, 0x91, 0x02, 0x42, 0x82, 0x12, 0x52, 0x92};
 
   const RawBitmapEx srcRawBitmap = RawBitmapEx::Create(SpanUtil::AsSpan(srcContent), Extent, SrcPixelFormat, Origin);
   const TightPlanarBitmapR8G8B8 result(srcRawBitmap);
@@ -263,8 +263,8 @@ TEST(Test_TightPlanarBitmapR8G8B8, ConstructFromBitmap)
   constexpr uint32_t ExpectedPlaneStride = Stride / 3;
   constexpr uint32_t ExpectedBytesPerPlane = ExpectedPlaneStride * Extent.Height.Value;
 
-  std::array<uint8_t, 3 * Extent.Width.Value * Extent.Height.Value> srcContent{0x00, 0x40, 0x80, 0x10, 0x50, 0x90, 0x01, 0x41, 0x81,
-                                                                               0x11, 0x51, 0x91, 0x02, 0x42, 0x82, 0x12, 0x52, 0x92};
+  std::array<uint8_t, static_cast<std::size_t>(3) * Extent.Width.Value * Extent.Height.Value> srcContent{
+    0x00, 0x40, 0x80, 0x10, 0x50, 0x90, 0x01, 0x41, 0x81, 0x11, 0x51, 0x91, 0x02, 0x42, 0x82, 0x12, 0x52, 0x92};
 
   const Bitmap srcBitmap(SpanUtil::AsSpan(srcContent), TypeConverter::To<PxSize2D>(Extent), SrcPixelFormat, Origin);
   const TightPlanarBitmapR8G8B8 result(srcBitmap);
@@ -293,8 +293,8 @@ TEST(Test_TightPlanarBitmapR8G8B8, ConstructFromBitmapMemory)
   constexpr uint32_t ExpectedPlaneStride = Stride / 3;
   constexpr uint32_t ExpectedBytesPerPlane = ExpectedPlaneStride * Extent.Height.Value;
 
-  std::array<uint8_t, 3 * Extent.Width.Value * Extent.Height.Value> srcContent{0x00, 0x40, 0x80, 0x10, 0x50, 0x90, 0x01, 0x41, 0x81,
-                                                                               0x11, 0x51, 0x91, 0x02, 0x42, 0x82, 0x12, 0x52, 0x92};
+  std::array<uint8_t, static_cast<std::size_t>(3) * Extent.Width.Value * Extent.Height.Value> srcContent{
+    0x00, 0x40, 0x80, 0x10, 0x50, 0x90, 0x01, 0x41, 0x81, 0x11, 0x51, 0x91, 0x02, 0x42, 0x82, 0x12, 0x52, 0x92};
 
   const TightPlanarBitmapR8G8B8 result(Bitmap(SpanUtil::AsSpan(srcContent), TypeConverter::To<PxSize2D>(Extent), SrcPixelFormat, Origin).Release());
 
@@ -323,8 +323,8 @@ TEST(Test_TightPlanarBitmapR8G8B8, ConstructFromTightBitmap)
   constexpr uint32_t ExpectedPlaneStride = Stride / 3;
   constexpr uint32_t ExpectedBytesPerPlane = ExpectedPlaneStride * Extent.Height.Value;
 
-  std::array<uint8_t, 3 * Extent.Width.Value * Extent.Height.Value> srcContent{0x00, 0x40, 0x80, 0x10, 0x50, 0x90, 0x01, 0x41, 0x81,
-                                                                               0x11, 0x51, 0x91, 0x02, 0x42, 0x82, 0x12, 0x52, 0x92};
+  std::array<uint8_t, static_cast<std::size_t>(3) * Extent.Width.Value * Extent.Height.Value> srcContent{
+    0x00, 0x40, 0x80, 0x10, 0x50, 0x90, 0x01, 0x41, 0x81, 0x11, 0x51, 0x91, 0x02, 0x42, 0x82, 0x12, 0x52, 0x92};
 
   const TightBitmap srcBitmap(SpanUtil::AsSpan(srcContent), Extent, SrcPixelFormat, Origin);
   const TightPlanarBitmapR8G8B8 result(srcBitmap);
